Don't run stty command if stdin is not a terminal
Fixed an issue with the stty command found in .bashrc dealing with times
when stdin is not a terminal (i.e. when scp'ing a files). If stdin was
not a terminal, the stty command would complain with the following
error:

	stty: standard input: Invalid argument

Thus, the added test ensures the command is only run when stdin is a
terminal.

See also: http://www.perlmonks.org/?node_id=534691